api-request-with-slider-processing remains

+++ lib/data/requests/loan_offer_requests/get_loan_offers_by_type_request.dart
@@ -16,6 +16,6 @@ final class GetLoanOffersRequest extends IRestApiRequest {
     authorize = false;
   }
   final LoanTypeEnum loanType;
-  final int loanAmount;
+  final double loanAmount;
   final int loanTerm;
 }
